tests: split tests into separate files by category
