+++ test/helpers/data.ts
@@ -1,0 +1,8 @@
+
+export const botHeaders = { 'User-Agent': 'Discordbot/2.0' };
+export const humanHeaders = {
+  'User-Agent':
+    'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/116.0.0.0 Safari/537.36'
+};
+export const twitterBaseUrl = 'https://x.com';
+export const githubUrl = 'https://github.com/FxEmbed/FxEmbed';
